Added check for all prices, not just grams
Originally, the code checked to see if all prices were empty except
"gram" and would display the per gram price, or nothing.

This was causing no price to show if only a half gram price was added,
or eight, quarter, half ounce, or ounce.

Now, the code checks to see if any of the prices are the only prices and
displays their price accordingly.